Modify directory hierarchy
