+++ src/main/java/fifo/controller/RootController.java
@@ -1,14 +1,14 @@
-package fifo.contents;
+package fifo.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping(value = "/statistics")
-public class StatisticsController {
+@RequestMapping(value = "/")
+public class RootController {
   @RequestMapping(method = RequestMethod.GET)
   public String index() {
-    return "contents/statistics/index";
+    return "root/index";
   }
 }
