+++ src/main/java/fifo/controller/contents/ExercisesController.java
@@ -1,14 +1,14 @@
-package fifo;
+package fifo.controller.contents;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping(value = "/")
-public class RootController {
+@RequestMapping(value = "/exercises")
+public class ExercisesController {
   @RequestMapping(method = RequestMethod.GET)
   public String index() {
-    return "root/index";
+    return "contents/exercises/index";
   }
 }
