+++ src/main/java/fifo/controller/contents/StatisticsController.java
@@ -1,14 +1,14 @@
-package fifo.contents;
+package fifo.controller.contents;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping(value = "/exercises")
-public class ExercisesController {
+@RequestMapping(value = "/statistics")
+public class StatisticsController {
   @RequestMapping(method = RequestMethod.GET)
   public String index() {
-    return "contents/exercises/index";
+    return "contents/statistics/index";
   }
 }
