+++ src/main/java/fifo/controller/contents/UserAuthController.java
@@ -1,4 +1,4 @@
-package fifo.contents;
+package fifo.controller.contents;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
