+++ src/main/java/fifo/controller/user/ArchivesController.java
@@ -1,4 +1,4 @@
-package fifo.contents.user;
+package fifo.controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.PathVariable;
 
 @Controller
-@RequestMapping(value = "/users/{userId}/settings")
-public class SettingsController {
+@RequestMapping(value = "/users/{userId}/archives")
+public class ArchivesController {
   @RequestMapping(method = RequestMethod.GET)
   public String index(@PathVariable("userId") String userId) {
-    return "contents/user/settings/index";
+    return "user/archives/index";
   }
 }
