+++ src/main/java/fifo/controller/user/CalendarController.java
@@ -1,4 +1,4 @@
-package fifo.contents.user;
+package fifo.controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,6 +10,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 public class CalendarController {
   @RequestMapping(method = RequestMethod.GET)
   public String index(@PathVariable("userId") String userId) {
-    return "contents/user/calendar/index";
+    return "user/calendar/index";
   }
 }
