+++ src/main/java/fifo/controller/user/UsersController.java
@@ -1,4 +1,4 @@
-package fifo.contents.user;
+package fifo.controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,11 +10,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 public class UsersController {
   @RequestMapping(value = "/{userId}", method = RequestMethod.GET)
   public String index(@PathVariable("userId") String userId) {
-    return "contents/user/users/index";
+    return "user/users/index";
   }
 
   @RequestMapping(value = "/new", method = RequestMethod.GET)
   public String index2() {
-    return "contents/user/users/index";
+    return "user/users/index";
   }
 }
